+++ zojae031/Project/app/src/main/java/study/architecture/data/Repository.kt
@@ -1,13 +1,11 @@
-package study.architecture.data.local
+package study.architecture.data
 
 import io.reactivex.Single
 import study.architecture.data.entity.Market
 import study.architecture.data.entity.Ticker
 
-interface UpbitLocalDataSource {
+interface Repository {
     fun getMarkets(): Single<List<Market>>
     fun getTickers(markets: String): Single<MutableList<Ticker>>
-    fun insertMarket(market: Market)
-    fun insertTicker(ticker: Ticker)
-
+    fun checkNetwork(): Boolean
 }
