+++ zojae031/Project/app/src/main/java/study/architecture/data/UpbitLocalDataSource.kt
@@ -1,11 +1,13 @@
-package study.architecture.data.repository
+package study.architecture.data
 
 import io.reactivex.Single
 import study.architecture.data.entity.Market
 import study.architecture.data.entity.Ticker
 
-interface Repository {
+interface UpbitLocalDataSource {
     fun getMarkets(): Single<List<Market>>
     fun getTickers(markets: String): Single<MutableList<Ticker>>
-    fun checkNetwork(): Boolean
+    fun insertMarket(market: Market)
+    fun insertTicker(ticker: Ticker)
+
 }
