+++ zojae031/Project/app/src/main/java/study/architecture/presentation/CoinAdapterContract.kt
@@ -1,4 +1,4 @@
-package study.architecture.presentation.coinjob.adapter
+package study.architecture.presentation
 
 import study.architecture.data.entity.ProcessingTicker
