+++ zojae031/Project/app/src/main/java/study/architecture/presentation/CoinPresenter.kt
@@ -1,4 +1,4 @@
-package study.architecture.presentation.coinjob
+package study.architecture.presentation
 
 import android.annotation.SuppressLint
 import io.reactivex.android.schedulers.AndroidSchedulers
@@ -6,7 +6,7 @@ import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import study.architecture.data.entity.ProcessingTicker
 import study.architecture.data.repository.RepositoryImpl
-import study.architecture.presentation.coinjob.adapter.CoinAdapterContract
+import study.architecture.ui.coinjob.CoinFragment
 import study.architecture.util.TextUtil
 import java.util.concurrent.TimeUnit
